[analyzer] Move unexecuted test block into it's own source file
Inside lambdas.cpp a block of code wasn't executed,
because it required the standard to be at least c++14.
This patch moves this block of code into it's own
source file and makes sure it's tested.

Differential Revision: https://reviews.llvm.org/D135965

(cherry picked from commit ea8aebf9eb7f0762d357e02524be9f65cfdb4f58)
